arc: parse german input files
The German text file turned out to be way more problematic than the
English one, and I wonder why it returns a different number of tags. I
had to restructure the code to make it work, but already have a lot of
ideas for improvements (that make more sense) that I can test now.